requirements.txt
requests
python-dateutil
pydantic
tenacity
pytest
logzio-python-handler
